requirement.txt
Flask
python-magic